some adjustements
