+++ initReactI18next.d.ts
@@ -1,0 +1,3 @@
+import { ThirdPartyModule } from 'i18next';
+
+export const initReactI18next: ThirdPartyModule;
