+++ package.json
@@ -28,6 +28,11 @@
       "types": "./TransWithoutContext.d.ts",
       "import": "./dist/es/TransWithoutContext.js",
       "require": "./dist/commonjs/TransWithoutContext.js"
+    },
+    "./initReactI18next": {
+      "types": "./initReactI18next.d.ts",
+      "import": "./dist/es/initReactI18next.js",
+      "require": "./dist/commonjs/initReactI18next.js"
     }
   },
   "homepage": "https://github.com/i18next/react-i18next",
