+++ src/context.js
@@ -1,9 +1,9 @@
 import { createContext } from 'react';
 import { getDefaults, setDefaults } from './defaults';
+import { getI18n, setI18n } from './i18nInstance';
+import { initReactI18next } from './initReactI18next';
 
-export { getDefaults, setDefaults };
-
-let i18nInstance;
+export { getDefaults, setDefaults, getI18n, setI18n, initReactI18next };
 
 export const I18nContext = createContext();
 
@@ -22,23 +22,6 @@ export class ReportNamespaces {
     return Object.keys(this.usedNamespaces);
   }
 }
-
-export function setI18n(instance) {
-  i18nInstance = instance;
-}
-
-export function getI18n() {
-  return i18nInstance;
-}
-
-export const initReactI18next = {
-  type: '3rdParty',
-
-  init(instance) {
-    setDefaults(instance.options.react);
-    setI18n(instance);
-  },
-};
 
 export function composeInitialProps(ForComponent) {
   return (ctx) =>
