+++ src/initReactI18next.js
@@ -1,0 +1,11 @@
+import { setDefaults } from './defaults';
+import { setI18n } from './i18nInstance';
+
+export const initReactI18next = {
+  type: '3rdParty',
+
+  init(instance) {
+    setDefaults(instance.options.react);
+    setI18n(instance);
+  },
+};
